Separate game-engine test files from production source
Move all *.test.ts out of src/ into a sibling test/ directory, so
src/ only ever contains production code. Requires splitting the
tsconfig: tsconfig.json (src only) still drives `build`'s dist
output, and a new tsconfig.test.json (src + test + scripts) drives
`lint`'s typecheck without emitting anything. Vitest already
discovers test/*.test.ts by default, no config change needed there.
