+++ packages/game-engine/test/mineExplosion.test.ts
@@ -1,8 +1,8 @@
 import { describe, expect, it } from "vitest";
-import { initialize } from "./gameControl.js";
-import { makeMove } from "./makeMove.js";
-import { getField } from "./grid.js";
-import type { Player } from "./types.js";
+import { initialize } from "../src/gameControl.js";
+import { makeMove } from "../src/makeMove.js";
+import { getField } from "../src/grid.js";
+import type { Player } from "../src/types.js";
 
 const noMines = () => 1;
